Add option to display prefixes
When using bgpq3 etc to generate prefix lists from as-sets it can be useful to dig through exactly where in the chain an unexpected prefix got included.

irrtree already pulls all the prefixes down to count them, so adding an option to display them appears to make sense.

-x (as in extra {info}) writes the prefixes into a sub-tree, causing a display similar to:

AS-GLOBAL (261 ASNs, 267 pfxs)
+-- AS-LAN (27 ASNs, 50 pfxs)
|   +-- AS-FINECOM (27 ASNs, 50 pfxs)
|       +-- AS-TELEZUG (6 ASNs, 5 pfxs)
|       |   +-- AS-SHINTERNET (2 ASNs, 3 pfxs)
|       |   |   +-- AS35518 (3 pfxs)
|       |   |   |   +-- 2a00:d80::/32
[...]