Setup: Add wheel as a requirement
When wheel package is not installed an error message will be printed to
stdout.

Part of the msg will contain this:

```
WARNING: The wheel package is not available.
WARNING: The wheel package is not available.
  error: subprocess-exited-with-error
```

Suggestion is to add wheel as a requirement in setup.py with the
existing requirements.